# src/app/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: simple-pos API
  version: 1.0.0-draft
  description: |
    Contract-first API baseline for MVP.
    Versioned under /api/v1 and aligned with workflow-manager planning artifacts.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Local API base
security:
  - bearerAuth: []
...
paths:
...
  /reports/top-products:
    get:
      tags: [Reporting]
      operationId: getTopProducts
      summary: Get top-selling products
      parameters:
        - in: query
          name: periodStart
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: periodEnd
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Top products report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopProductsResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /reports/profit-summary:
    get:
      tags: [Reporting]
      operationId: getProfitSummary
      summary: Get revenue/cost/profit summary
      parameters:
        - in: query
          name: periodStart
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: periodEnd
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Profit summary report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfitSummaryResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /reports/sales-history:
    get:
      tags: [Reporting]
      operationId: getSalesHistory
      summary: List sales history with date and payment filters
      parameters:
        - in: query
          name: periodStart
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: periodEnd
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: paymentMethod
          required: false
          schema:
            type: string
            enum: [cash, on_account]
      responses:
        '200':
          description: Sales history report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesHistoryResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Standardized API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    TopProductItem:
      type: object
      required: [productId, name, quantitySold, revenue]
      properties:
        productId:
          type: string
        name:
          type: string
        quantitySold:
          type: number
        revenue:
          type: number
    TopProductsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TopProductItem'
    ProfitSummaryResponse:
      type: object
      required: [revenue, cost, profit]
      properties:
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
    SalesHistoryItem:
      type: object
      required: [saleId, paymentMethod, total, itemCount, createdAt]
      properties:
        saleId:
          type: string
        paymentMethod:
          type: string
          enum: [cash, on_account]
        customerId:
          type: [string, 'null']
        total:
          type: number
        itemCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    SalesHistoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesHistoryItem'
...
    ValidationErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'